Split out anyvalue conversion
